+++ sdk/@launchdarkly/observability-android/CHANGELOG.md
@@ -1,5 +1,20 @@
 # Changelog
 
+## [0.28.0](https://github.com/launchdarkly/observability-sdk/compare/launchdarkly-observability-android-0.27.0...launchdarkly-observability-android-0.28.0) (2026-03-02)
+
+
+### Features
+
+* Make Android SDK35 compilable ([#405](https://github.com/launchdarkly/observability-sdk/issues/405)) ([278880d](https://github.com/launchdarkly/observability-sdk/commit/278880dae408afe304e35c00ebe989e466f509eb))
+
+## [0.27.0](https://github.com/launchdarkly/observability-sdk/compare/launchdarkly-observability-android-0.26.1...launchdarkly-observability-android-0.27.0) (2026-02-28)
+
+
+### Features
+
+* Android Incremental Image Diff compression ([#390](https://github.com/launchdarkly/observability-sdk/issues/390)) ([5ff93f6](https://github.com/launchdarkly/observability-sdk/commit/5ff93f61e078b27ca7d15a49e35ac98076d684cc))
+* Optional Jet Compose ([#402](https://github.com/launchdarkly/observability-sdk/issues/402)) ([8f3a671](https://github.com/launchdarkly/observability-sdk/commit/8f3a671a7853353b3e255882b66c6cb04506dcaa))
+
 ## [0.26.1](https://github.com/launchdarkly/observability-sdk/compare/launchdarkly-observability-android-0.26.0...launchdarkly-observability-android-0.26.1) (2026-02-24)